% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataraw.R
\docType{data}
\name{DSraw}
\alias{DSraw}
\title{Data Scientists/Data Analyst/ Statistician Job Tracker}
\format{
A data frame with 551 rows and 152 variables
\describe{
\item{ID}{row id}
\item{Consultant}{Name of the consultant}
}
}
\source{
Collected and entered by BSc (Hons) Statistics undegraduates - 2020
}
\usage{
DSraw
}
\description{
Job advertisements
}
\examples{
data(DSraw)
head(DSraw)
summary(DSraw)
}
\keyword{datasets}
