% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatidy.R
\docType{data}
\name{DStidy}
\alias{DStidy}
\title{Data Scientists/Data Analyst/ Statistician Job Tracker}
\format{
A data frame with 812 rows and 109 columns
\describe{
\item{ID}{row id}
\item{Consultant}{Name of the consultant}
\item{DateRetrieved}{ Date of Data Retrieved }
\item{DatePublished}{Published Date of the Advertisement  }
\item{Job_title}{ Name of the job category}
\item{Company}{Name of the Company}
\item{R}{If R is required -> 1 ,If not mentioned -> 0}
\item{SAS}{}
\item{SPSS}{}
\item{Python}{}
\item{MAtlab}{}
\item{Scala}{}
\item{C#}{}
\item{MS Word}{}
\item{Ms Excel}{}
\item{OLE/DB}{}
\item{Ms Access}{}
\item{Ms PowerPoint}{}
\item{Spreadsheets}{}
\item{Data_visualization}{}
\item{Presentation_Skills}{}
\item{Communication}{}
\item{BigData}{}
\item{Data_warehouse}{}
\item{cloud_storage}{}
\item{Google_Cloud}{}
\item{AWS}{}
\item{Machine_Learning}{}
\item{Deep Learning}{}
\item{Computer_vision}{}
\item{Java}{}
\item{C++}{}
\item{C}{}
\item{Linux/Unix}{}
\item{SQL}{}
\item{NoSQL}{}
\item{RDBMS}{}
\item{Oracle}{}
\item{MySQL}{}
\item{PHP}{}
\item{SPL}{}
\item{web_design_and_development_tools}{}
\item{AI}{}
\item{Natural_Language_Processing(NLP)}{}
\item{Microsoft Power BI}{}
\item{Google_Analytics}{}
\item{graphics_and_design_skills}{}
\item{Data_marketing}{}
\item{SEO}{}
\item{Content_Management}{}
\item{Tableau}{}
\item{D3}{}
\item{Alteryx}{}
\item{KNIME}{}
\item{Spotfire}{}
\item{Spark}{}
\item{S3}{}
\item{Redshift}{}
\item{DigitalOcean}{}
\item{Javascript}{}
\item{Kafka}{}
\item{Storm}{}
\item{Bash}{}
\item{Hadoop}{}
\item{Data_Pipelines}{}
\item{MPP_Platforms}{}
\item{Qlik}{}
\item{Pig}{}
\item{Hive}{}
\item{Tensorflow}{}
\item{Map/Reduce}{}
\item{Impala}{}
\item{Solr}{}
\item{Teradata}{}
\item{MongoDB}{}
\item{Elasticsearch}{}
\item{YOLO}{}
\item{agile execution}{}
\item{Data_management}{}
\item{pyspark}{}
\item{Data_mining}{}
\item{Data_science}{}
\item{Web_Analytic_tools}{}
\item{IOT}{}
\item{Numerical_Analysis}{}
\item{Economic}{}
\item{Finance_Knowledge}{}
\item{Investment_Knowledge}{}
\item{Problem_Solving}{}
\item{Team_Handling}{}
\item{Debtor_reconcilation}{}
\item{Payroll_management}{}
\item{Bayesian}{}
\item{Optimization}{}
\item{Knowledge_in}{}
\item{City}{}
\item{Educational_qualifications}{}
\item{Salary}{}
\item{URL}{}
\item{Search_Term}{}
\item{Job_Category}{}
\item{Experience_Category}{}
\item{Location}{}
\item{Payment Frequency}{}
\item{BSc_needed}{}
\item{MSc_needed}{}
\item{PhD_needed}{}
\item{English Needed}{}
\item{year}{}
}
}
\source{
Data collected and combined by the Statstical Consulting Service of University of Sri Jayewardenepura
}
\usage{
DStidy
}
\description{
Combined collection of Job advertisements collected by the Statstical Consulting Service of the University of Sri Jayewardenepura
}
\keyword{datasets}
