% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatidy.R
\docType{data}
\name{DStidy_2021}
\alias{DStidy_2021}
\title{Data Scientists/Data Analyst/ Statistician Job Advertisements in the year 2021}
\format{
A data frame with 382 rows and 112 columns
\describe{
\item{ID}{row id}
\item{Consultant}{Name of the consultant}
\item{URL}{}
\item{Search_Term}{}
\item{DateRetrieved}{ Date of Data Retrieved }
\item{DatePublished}{Published Date of the Advertisement}
\item{Job_Field}{}
\item{Job_title}{}
\item{Company}{}
\item{Knowledge_in}{}
\item{Minimum Experience in Years}{}
\item{City}{}
\item{Location}{}
\item{Educational_qualifications}{}
\item{Payment Frequency}{}
\item{Currency}{}
\item{Salary}{}
\item{English Needed}{}
\item{English proficiency description}{}
\item{Additional_languages}{}
\item{AI}{}
\item{Natural_Language_Processing(NLP)}{}
\item{Data_Pipelines}{}
\item{Machine_Learning}{}
\item{Deep Learning}{}
\item{Computer_vision}{}
\item{Data_visualization}{}
\item{Data_warehouse}{}
\item{BigData}{}
\item{Data_management}{}
\item{Data_mining}{}
\item{Data_science}{}
\item{Bayesian}{}
\item{Optimization}{}
\item{Numerical_Analysis}{}
\item{IOT}{}
\item{Data_translation}{}
\item{R}{}
\item{SAS}{}
\item{SPSS}{}
\item{Python}{}
\item{MAtlab}{}
\item{Scala}{}
\item{C#}{}
\item{Java}{}
\item{C++}{}
\item{C}{}
\item{Bash}{}
\item{Tensorflow}{}
\item{pyspark}{}
\item{YOLO}{}
\item{MS Word}{}
\item{Ms Excel}{}
\item{Ms Access}{}
\item{Ms PowerPoint}{}
\item{Spreadsheets}{}
\item{Google_Analytics}{}
\item{Microsoft Power BI}{}
\item{Tableau}{}
\item{D3}{}
\item{Qlik}{}
\item{KNIME}{}
\item{Spotfire}{}
\item{Linux/Unix}{}
\item{OLE/DB}{}
\item{SQL}{}
\item{NoSQL}{}
\item{RDBMS}{}
\item{Oracle}{}
\item{MySQL}{}
\item{MongoDB}{}
\item{MPP_Platforms}{}
\item{SPL}{}
\item{Alteryx}{}
\item{Spark}{}
\item{Kafka}{}
\item{Hadoop}{}
\item{Pig}{}
\item{Hive}{}
\item{Map/Reduce}{}
\item{Impala}{}
\item{Storm}{}
\item{Google_Cloud}{}
\item{AWS}{}
\item{cloud_storage}{}
\item{S3}{}
\item{Redshift}{}
\item{DigitalOcean}{}
\item{Teradata}{}
\item{Solr}{}
\item{Elasticsearch}{}
\item{Presentation_Skills}{}
\item{Communication}{}
\item{Problem_Solving}{}
\item{Team_Handling}{}
\item{agile execution}{}
\item{Data_marketing}{}
\item{SEO}{}
\item{graphics_and_design_skills}{}
\item{Content_Management}{}
\item{Economic}{}
\item{Finance_Knowledge}{}
\item{Investment_Knowledge}{}
\item{Debtor_reconcilation}{}
\item{Payroll_management}{}
\item{web_design_and_development_tools}{}
\item{PHP}{}
\item{Javascript}{}
\item{Web_Analytic_tools}{}
\item{BSc_needed}{}
\item{MSc_needed}{}
\item{PhD_needed}{}
}
}
\source{
Data was sourced and wrangled by the BSc.Statistics Honors batch of 2016/2017
}
\usage{
DStidy_2021
}
\description{
Job advertisements collected in the year 2021
}
\examples{
data(DStidy_2021)
head(DStidy_2021)
summary(DStidy_2021)
}
\keyword{datasets}
