% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get data from DSjobtracker for specific years or all the years combined into one dataset}
\usage{
get_data(year)
}
\arguments{
\item{year}{can be either "all" or an year after 2020 (2020,2021,...,etc.) as a numeric value}
}
\description{
The DSjobtracker dataset is updated each year through the Statistical Consultancy Service of University of Sri Jayewardenepura. In order to accommodate the structural changes of data this function provides the capability to get the dataset required either combined through out the years or data specific to each year.
}
