% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_dashboard.R
\name{open_dashboard}
\alias{open_dashboard}
\title{Opens the dashboard containing visualizations of the dataset}
\usage{
open_dashboard()
}
\description{
Opens the dashboard containing visualizations of the dataset
}
