% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{DSraw}
\alias{DSraw}
\title{Data Scientists/Data Analyst/ Statistician Job Tracker}
\format{}
\source{
Collected and entered by BSc (Hons) Statistics undegraduates - 2020
}
\usage{
DSraw
}
\description{
Job advertisements
}
\examples{
data(DSraw)
head(Dsraw)
summary(DSraw)
}
\keyword{datasets}
