% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datatidy.R
\docType{data}
\name{DStidy}
\alias{DStidy}
\title{Data Scientists/Data Analyst/ Statistician Job Tracker}
\format{}
\source{
Data wrangling part was done by Chammika R. De Mel, Shashini D. Silva and Thimani D Ranathunga, BSc (Hons)Staistics, University of Sri
Jayewardenepura.
#' @examples
data(DStidy)
head(DStidy)
summary(DStidy)
}
\usage{
DStidy
}
\description{
Job advertisements
}
\keyword{datasets}
