% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_dashboard.R
\name{open_dashboard}
\alias{open_dashboard}
\title{Visualize the dataset in a dashboard}
\usage{
open_dashboard()
}
\description{
{
Opens the dashboard containing visualizations of the dataset. \cr

The dashboard is hosted at \url{https://scs-fas-sjp.netlify.app/dsjobtrackervisualisation/dataviz}
}
}
\details{
The data wrangling dashboard was created by Jayani Lakshika
}
\examples{
open_dashboard()
}
